requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
distro==1.9.0
docker==7.1.0
geojson==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.3.0
pydantic==2.12.0
pydantic_core==2.41.1
pygame==2.6.1
pyowm==3.5.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
